server/requirements.txt
pandas==2.3.3
fastapi==0.121.3
uvicorn==0.38.0
supabase==2.24.0
python-dotenv==1.2.1
shapely==2.1.2